fixed float for health

+++ UnityProject/Assets/Scripts/Health/Living/PlayerHealth.cs
@@ -53,7 +53,7 @@ public class PlayerHealth : HealthBehaviour
 		base.OnStartClient();
 	}
 
-	public override int ReceiveAndCalculateDamage(GameObject damagedBy, int damage, DamageType damageType,
+	public override float ReceiveAndCalculateDamage(GameObject damagedBy, float damage, DamageType damageType,
 		BodyPartType bodyPartAim)
 	{
 		base.ReceiveAndCalculateDamage(damagedBy, damage, damageType, bodyPartAim);
